requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
pillow
python-dotenv
